requirements.txt
# LiveKit dependencies
livekit>=0.18.3
livekit-agents>=0.1.0

# Core dependencies
python-dotenv>=1.0.0
functions-framework==3.4.0
aiohttp~=3.10.0
websockets~=12.0

# AI and Weather dependencies
google-generativeai>=0.3.1
requests>=2.31.0
python-weather>=1.1.1
asyncio>=3.4.3